Update
+ New function in main to run experiment on all three trees with the
same numbers, to make the results more accurate (Commented out the calls
in main until the modified code is done)

#The binExperiment and digExperiment functions will be redundant and can
be removed once this is finished.

+ modDigitalTree.java : I don't think this class actually needs any
work, it may not even be necessary actually, since this is basically
just a digital tree with a different input, so we could probably just
make a second digital tree with the correct value inserted.